requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
distro==1.9.0
fastapi==0.135.1
groq==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.32.5
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0